python_requirements.txt
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
boto3==1.24.5
botocore==1.27.5
certifi==2021.10.8
cffi==1.15.0
cryptography==37.0.2
click==8.1.3
google-cloud-storage==2.5.0
google-api-core==2.8.2
google-auth==2.10.0
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.47.0
grpcio-health-checking==1.47.0
grpcio-status==1.46.3
grpcio-tools==1.46.3
imutils==0.5.4
jmespath==1.0.0
minio==7.1.8
mlflow==1.28.0
numpy==1.22.3
opencv-python==4.5.5.64
ordereddict==1.1
pydantic==1.10.7
Pillow==9.1.1
protobuf==3.20.1
pycparser==2.21
python-dateutil==2.8.2
PyYAML==6.0
psycopg2==2.9.3
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.1.2
s3transfer==0.6.0
scipy==1.8.0
simplejson==3.17.6
starlette==0.27.0
six==1.16.0
urllib3==1.26.9
aiohttp==3.8.2
jsonschema==4.16.0
uvicorn==0.22.0
uvloop==0.17.0
fastapi==0.96.0
fastjsonschema==2.16.2
psycopg==3.1.9
psycopg-pool==3.1.7